+++ src/routes/api/Auth.ts
@@ -1,8 +1,8 @@
 import * as express from 'express';
 
-import Auth from '../controllers/Auth';
-import { asyncController } from '../plugins/utility';
-import AuthUser from '../validator/Auth';
+import Auth from '../../controllers/Auth';
+import { asyncController } from '../../plugins/utility';
+import AuthUser from '../../validator/Auth';
 
 const router = express.Router();
 const { login, logout } = Auth;
