+++ src/routes/api/Session.ts
@@ -1,8 +1,8 @@
 import * as express from 'express';
 
-import SessionController from '../controllers/Session';
-import { asyncController } from '../plugins/utility';
-import SessionValidator from '../validator/util';
+import SessionController from '../../controllers/Session';
+import { asyncController } from '../../plugins/utility';
+import SessionValidator from '../../validator/util';
 
 const router = express.Router();
 const { post, get, put, destroy } = SessionController;
