+++ src/routes/index.ts
@@ -1,16 +1,9 @@
 import * as express from 'express';
-import AuthRouter from './Auth';
-import SessionRouter from './Session';
-import StockRouter from './Stock';
-import TransactionRouter from './Transaction';
-import UserRouter from './User';
+import ApiRouter from './api';
 
 const router = express.Router();
 
-router.use('/users', UserRouter);
-router.use('/sessions', SessionRouter);
-router.use('/auth', AuthRouter);
-router.use('/stocks', StockRouter);
-router.use('/transactions', TransactionRouter);
+/* interface for Rest full api */
+router.use('/api', ApiRouter);
 
 export default router;
